% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convertMSigDB}
\alias{convertMSigDB}
\title{convert ids of MSigDB to another organism}
\usage{
convertMSigDB(msigdbPath, hcopPath, idType = "ensembl")
}
\arguments{
\item{msigdbPath}{path to the MSigDB file}

\item{idType}{'entrez' or 'ensembl'}

\item{hcopPaht}{path to the hcop annotation file}
}
\description{
This function gets two files: one is downloaded from the MSigDB, it is .gmt file, the other is a .txt file downloaded form the HCOP database.
}
