% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{getHCOP}
\alias{getHCOP}
\title{get the latest data form the HCOP website}
\usage{
getHCOP(species = "mouse", saveto)
}
\arguments{
\item{species}{}

\item{saveto}{}
}
\description{
get the latest data form the HCOP website
}
